Intermediate perl script parse through blast output and cleaned read files
The purpose of this is to take a blast output and parse the information
for a best hit(s) determined by evalue and percent identity, print the
output. For this, once the blast output is parsed, only the name taken
and compared to an origianl source file to extract name and sequence
information from a cleaned Illumina sequence file.